refactor: split editor listeners into files
