+++ src/editor/listeners/editorTextUpdate.listener.ts
@@ -5,60 +5,18 @@ import {isArray} from 'lodash';
 import {join} from 'path';
 
 import {AppListenerFn} from '@redux/listeners/base';
-import {processResourceRefs} from '@redux/parsing/parser.thunks';
-import {selectFile, selectResource, updateMultipleClusterResources} from '@redux/reducers/main';
+import {updateMultipleClusterResources} from '@redux/reducers/main';
 import {selectedFilePathSelector} from '@redux/selectors';
 import {getResourceContentFromState, getResourceMetaFromState} from '@redux/selectors/resourceGetters';
 import {editorResourceIdentifierSelector} from '@redux/selectors/resourceSelectors';
 import {multiplePathsChanged} from '@redux/thunks/multiplePathsChanged';
 import {updateFileEntry} from '@redux/thunks/updateFileEntry';
 import {updateResource} from '@redux/thunks/updateResource';
-import {validateResources} from '@redux/validation/validation.thunks';
 
+import {getEditor} from '@editor/editor.instance';
 import {ROOT_FILE_ENTRY} from '@shared/constants/fileEntry';
 import {isLocalResourceMeta} from '@shared/models/k8sResource';
 import {isEqual} from '@shared/utils/isEqual';
-
-import {getEditor, recreateEditorModel, resetEditor} from './editor.instance';
-import {editorEnhancers} from './enhancers';
-import {applyEditorRefs} from './enhancers/k8sResource/refs';
-import {applyEditorValidation} from './enhancers/k8sResource/validation';
-
-export const editorSelectionListener: AppListenerFn = listen => {
-  listen({
-    matcher: isAnyOf(selectResource, selectFile),
-    async effect(_action, {getState, delay, dispatch, cancelActiveListeners}) {
-      cancelActiveListeners();
-      await delay(1);
-      const rootFolderPath = getState().main.fileMap[ROOT_FILE_ENTRY].filePath;
-      const editor = getEditor();
-
-      if (!editor || !rootFolderPath) {
-        return;
-      }
-
-      const resourceIdentifier = editorResourceIdentifierSelector(getState());
-      if (resourceIdentifier) {
-        const resourceContent = getResourceContentFromState(getState(), resourceIdentifier);
-        if (!resourceContent) {
-          return;
-        }
-        recreateEditorModel(editor, resourceContent.text);
-
-        const promises = editorEnhancers.map(enhancer =>
-          Promise.resolve(enhancer({state: getState(), editor, resourceIdentifier, dispatch}))
-        );
-        await Promise.all(promises);
-      }
-
-      if (!resourceIdentifier && isAnyOf(selectFile)(_action)) {
-        const selectedFilePath = _action.payload.filePath;
-        const fileText = await readFile(join(rootFolderPath, selectedFilePath), 'utf8');
-        recreateEditorModel(editor, fileText);
-      }
-    },
-  });
-};
 
 export const editorTextUpdateListener: AppListenerFn = listen => {
   listen({
@@ -163,46 +121,3 @@ export const editorTextUpdateListener: AppListenerFn = listen => {
     },
   });
 };
-
-export const editorResourceRefsListener: AppListenerFn = listen => {
-  listen({
-    matcher: isAnyOf(processResourceRefs.fulfilled),
-    async effect(_action, {getState, delay, cancelActiveListeners, dispatch}) {
-      cancelActiveListeners();
-      await delay(1);
-
-      const editorResourceIdentifier = editorResourceIdentifierSelector(getState());
-      resetEditor();
-
-      const editor = getEditor();
-      if (!editor) {
-        return;
-      }
-      applyEditorRefs({state: getState(), resourceIdentifier: editorResourceIdentifier, editor, dispatch});
-    },
-  });
-};
-
-export const editorValidationListener: AppListenerFn = listen => {
-  listen({
-    matcher: isAnyOf(validateResources.fulfilled),
-    async effect(_action, {getState, delay, cancelActiveListeners, dispatch}) {
-      cancelActiveListeners();
-      await delay(1);
-
-      const editorResourceIdentifier = editorResourceIdentifierSelector(getState());
-      const editor = getEditor();
-      if (!editor) {
-        return;
-      }
-      applyEditorValidation({state: getState(), resourceIdentifier: editorResourceIdentifier, editor, dispatch});
-    },
-  });
-};
-
-export const editorListeners: AppListenerFn[] = [
-  editorSelectionListener,
-  editorTextUpdateListener,
-  editorResourceRefsListener,
-  editorValidationListener,
-];
